Add examples to mkdocs.yml and resolve issues.
