+++ docs/src/webknossos-py/examples/create_dataset_from_images.md
@@ -1,0 +1,9 @@
+# Create Dataset from Images
+
+This example shows how to [convert a folder of images into a new dataset](../../api/webknossos/dataset/dataset.md#Dataset.from_images).
+
+```python
+--8<--
+webknossos/examples/create_dataset_from_images.py
+--8<--
+```
