--- docs/src/webknossos-py/examples/image_list_to_wkw.md
@@ -1,9 +1,0 @@
-# Convert image list to wkw file
-
-This example shows how to [create a new WEBKNOSSOS dataset from a stack of images](../../api/webknossos/dataset/dataset.md#Dataset.from_images), e.g. Tiff, JPEG, etc files.
-
-```python
---8<--
-webknossos/examples/image_list_to_wkw.py
---8<--
-```
