--- docs/src/webknossos-py/examples/write_images_to_harddrive.md
@@ -1,9 +1,0 @@
-# Write images to harddrive
-
-This example shows how to [convert a folder of images into a new dataset](../../api/webknossos/dataset/dataset.md#Dataset.from_images). To achive the correct order when images are stacked on another the z-order can be defined with a function.
-
-```python
---8<--
-webknossos/examples/write_images_to_harddrive.py
---8<--
-```
