+++ docs/src/webknossos-py/installation.md
@@ -11,7 +11,7 @@ You can install it from [pypi](https://pypi.org/project/webknossos/), e.g. via p
 pip install webknossos
 ```
 
-For some usecases it is necsessary to install the optional dependencies:
+For extended file format conversation support it is necessary to install the optional dependencies:
 
 ```bash
 pip install "webknossos[all]"
