+++ webknossos/examples/image_stack_to_dataset.py
@@ -1,0 +1,31 @@
+from pathlib import Path
+
+from webknossos import Dataset
+from webknossos.dataset import COLOR_CATEGORY
+
+INPUT_FILES = (
+    Path("webknossos/testdata/tiff/test.0000.tiff"),
+    Path("webknossos/testdata/tiff/test.0001.tiff"),
+)
+OUTPUT_FOLDER = Path("../example_output/tiff_dataset")
+
+
+def main() -> None:
+    """Convert a list of images into a WEBKNOSSOS dataset and directly add them as a new layer."""
+    dataset = Dataset(
+        dataset_path=OUTPUT_FOLDER,
+        voxel_size=(11, 11, 11),
+        name="My_new_dataset",
+        exist_ok=False,
+    )
+    dataset.add_layer_from_images(
+        images=INPUT_FILES,
+        layer_name="test",
+        category=COLOR_CATEGORY,
+    )
+
+    # dataset.upload()
+
+
+if __name__ == "__main__":
+    main()
